src/requirements.txt
astroid==2.3.3
awscli==1.16.309
boto3==1.14.26
botocore==1.17.26
colorama==0.4.1
docutils==0.15.2
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.18
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.2.0
protobuf==3.11.2
pyasn1==0.4.8
pylint==2.4.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.3.3
six==1.15.0
tabulate==0.8.7
typed-ast==1.4.0
urllib3==1.25.10
wrapt==1.11.2
